requirements.txt
redis
pytest
python-dotenv